Split readers into separate files
